+++ tests/assets/class5.ts
@@ -1,0 +1,9 @@
+export class OtherClass<P, S = {}> {}
+
+export interface OtherInterface<P> {}
+
+export class SomeClass extends OtherClass<{}> implements OtherInterface<{}> {
+  public constructor() {
+    super();
+  }
+}